Fix video player compile errors: missing isVideo import and mixed minOf argument types

+++ app/src/main/java/me/zhanghai/android/files/viewer/video/VideoPlayerActivity.kt
@@ -284,7 +284,7 @@ class VideoPlayerActivity : AppActivity() {
             return
         }
         val scale = minOf(
-            viewWidth / videoSarScaledWidth, viewHeight / videoHeight
+            viewWidth / videoSarScaledWidth, viewHeight.toFloat() / videoHeight
         )
         val matrix = Matrix()
         matrix.setScale(scale, scale, viewWidth / 2f, viewHeight / 2f)
